requirements.txt
huggingface-hub
strands-agents
strands-agents-tools
requests
fastapi
transformers
diffusers
accelerate
# starlette_context # for windows or linux